refactoring: move useTab.js to materialUIHooks folder
