+++ frontend/guest-app/src/components/Question/QuestionContainer.js
@@ -2,7 +2,7 @@ import React, {useEffect, useReducer, useRef} from "react";
 import {useQuery} from "@apollo/react-hooks";
 import {gql} from "apollo-boost";
 import QuestionContainerHeader from "./QuestionContainerHeader.js";
-import useTabs from "../TabGroup/useTabs.js";
+import useTabs from "../materialUIHooks/useTabs.js";
 import QuestionInputArea from "./QuestionInputArea/QuestionInputArea.js";
 import QuestionCardList from "./QuestionCardList.js";
 import {socketClient, useSocket} from "../../libs/socket.io-Client-wrapper.js";
